rename decoder to solidity-decoder
