+++ remix-debug/src/solidity-decoder/index.js
@@ -1,0 +1,11 @@
+var SolidityProxy = require('./solidityProxy')
+var stateDecoder = require('./stateDecoder')
+var localDecoder = require('./localDecoder')
+var InternalCallTree = require('./internalCallTree')
+
+module.exports = {
+  SolidityProxy: SolidityProxy,
+  stateDecoder: stateDecoder,
+  localDecoder: localDecoder,
+  InternalCallTree: InternalCallTree
+}
