+++ remix-debug/src/solidity-decoder/types/Bool.js
