+++ remix-debug/src/solidity-decoder/types/StringType.js
