+++ remix-debug/test/decoder/localsTests/misc.js
@@ -2,8 +2,8 @@
 var vmCall = require('../vmCall')
 var remixLib = require('remix-lib')
 var traceHelper = remixLib.helpers.trace
-var SolidityProxy = require('../../../src/decoder/solidityProxy')
-var InternalCallTree = require('../../../src/decoder/internalCallTree')
+var SolidityProxy = require('../../../src/solidity-decoder/solidityProxy')
+var InternalCallTree = require('../../../src/solidity-decoder/internalCallTree')
 var EventManager = remixLib.EventManager
 var helper = require('./helper')
