+++ remix-debug/test/decoder/storageDecoder.js
@@ -1,7 +1,7 @@
 'use strict'
 var tape = require('tape')
 var compiler = require('solc')
-var stateDecoder = require('../../src/decoder/stateDecoder')
+var stateDecoder = require('../../src/solidity-decoder/stateDecoder')
 var MockStorageResolver = require('./mockStorageResolver')
 var remixLib = require('remix-lib')
 var compilerInput = remixLib.helpers.compiler.compilerInput
